+++ src/types/vuetorrent/RightClickProperties.ts
@@ -1,0 +1,4 @@
+export default interface RightClickProperties {
+  isVisible: boolean,
+  offset: [number, number]
+}
